% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_v.R
\name{percentage_v}
\alias{percentage_v}
\title{Percentage V}
\usage{
percentage_v(df, v_label, utterance_id, cv_duration, utterance_duration)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, cv_duration, and utterance_duration values.}

\item{v_label}{a string to filter the vowels, e.g. \code{v_label = 'vowel'}}

\item{utterance_id}{column name for unique utterance IDs}

\item{cv_duration}{column name for the duration of C or V}

\item{utterance_duration}{column name for the duration of entire utterances}
}
\value{
a data frame containing the mean values and standard deviation of \%V
}
\description{
This function loads a dataframe as input and returns a dataframe containing the mean values and mean standard deviations of \%V.
}
\details{
\%V is a rhythm metrics based on Ramus, F., Nespor, M., & Mehler, J. (1999). Correlates of linguistic rhythm in the speech signal. Cognition, 73(3), 265-292. It calculates the percentage of an utterance occupied by vocalic material.

\verb{\%V: (total V duration / total utterance duration) * 100}
}
\examples{
df_test <- data.frame(cv_label = c("consonant", "vowel", "consonant", "vowel"),
                      utterance_id = c("utt_1", "utt_1", "utt_2", "utt_2"),
                      cv_duration = c(0.1, 0.8, 0.2, 0.5),
                      utterance_duration = c(2.4, 2.4, 2.7, 2.7))

percentage_v(df_test, v_label="vowel", utterance_id, cv_duration, utterance_duration)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
