% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rpvi_c.R
\name{plot_rpvi}
\alias{plot_rpvi}
\title{Plot rPVI C}
\usage{
plot_rpvi(
  df,
  cv_label,
  label_name,
  utterance_id,
  cv_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{cv_label}{column name for the segment labels (e.g., cv_label).}

\item{label_name}{a string to filter the segments, e.g. \code{label_name = 'consonant'}.}

\item{utterance_id}{column name for unique utterance IDs.}

\item{cv_duration}{column name for the duration of C or V.}

\item{save_fig}{default is \code{FALSE}. Change to \code{TRUE} to save the plot.}

\item{fig_path}{default is \code{NULL}. Required if \code{save_fig = TRUE}.}
}
\value{
A boxplot for rPVI values.
}
\description{
This function loads a dataframe as input and returns a box plot for rPVI C values.
}
\examples{
df_test <- data.frame(
  cv_label = rep(c("consonant", "vowel"), 8),
  utterance_id = rep(paste0("utt_", 1:4), each = 4),
  cv_duration = c(
    0.10, 0.80, 0.20, 0.50,
    0.30, 0.30, 0.40, 0.70,
    0.30, 0.88, 0.50, 0.90,
    0.30, 0.57, 0.40, 0.97
  )
)

plot_rpvi(
  df_test,
  cv_label = cv_label,
  label_name = "consonant",
  utterance_id = utterance_id,
  cv_duration = cv_duration
)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
