% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varco_cv.R
\name{plot_varco_cv}
\alias{plot_varco_cv}
\title{Plot Varco C & Varco V}
\usage{
plot_varco_cv(
  df,
  cv_label,
  utterance_id,
  cv_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{cv_label}{column name for segment type (C or V).}

\item{utterance_id}{column name for unique utterance IDs.}

\item{cv_duration}{column name for the duration of C or V.}

\item{save_fig}{default is \code{FALSE}. Change to \code{TRUE} to save the plot.}

\item{fig_path}{default is \code{NULL}. Required if \code{save_fig = TRUE}.}
}
\value{
A boxplot for Varco C and Varco V values.
}
\description{
This function loads a dataframe as input and returns a box plot for Varco C and Varco V values.
}
\examples{
df_test <- data.frame(
  cv_label = rep(c("consonant", "vowel"), 8),
  utterance_id = rep(paste0("utt_", 1:4), each = 4),
  cv_duration = c(
    0.10, 0.80, 0.20, 0.50,
    0.30, 0.30, 0.40, 0.70,
    0.30, 0.88, 0.50, 0.90,
    0.30, 0.57, 0.40, 0.97
  )
)

plot_varco_cv(
  df_test, cv_label, utterance_id, cv_duration, save_fig = FALSE
)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
