% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npvi_v.R
\name{npvi_v}
\alias{npvi_v}
\title{nPVI V}
\usage{
npvi_v(df, cv_label, label_name, utterance_id, cv_duration)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{cv_label}{column name for the segment labels (e.g., cv_label).}

\item{label_name}{a string to filter the vowels, e.g. \code{label_name = 'vowel'}.}

\item{utterance_id}{column name for unique utterance IDs.}

\item{cv_duration}{column name for the duration of C or V.}
}
\value{
a data frame containing the mean nPVI for vowels.
}
\description{
This function loads a dataframe as input and returns a dataframe containing the mean values of nPVI of vowels.
}
\examples{
df_test <- data.frame(cv_label = rep(c("consonant", "vowel"), 8),
                      utterance_id = rep(c("utt_1", "utt_2", "utt_3", "utt_4"), each = 4),
                      cv_duration = c(0.1, 0.8, 0.2, 0.5, 0.3, 0.3, 0.4, 0.7,
                                       0.3, 0.88, 0.5, 0.9, 0.3, 0.57, 0.4, 0.97))

npvi_v(df_test, cv_label = cv_label, label_name = "vowel", 
       utterance_id = utterance_id, cv_duration = cv_duration)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
