% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_v.R
\name{percentage_v}
\alias{percentage_v}
\title{Percentage V}
\usage{
percentage_v(df, v_label, utterance_id, cv_duration, utterance_duration)
}
\arguments{
\item{df:}{a data frame containing cv_labels, utterance_id, cv_duration, and utterance_duration values.}

\item{v_label:}{a string to filter the vowels, e.g. \code{v_label = 'V'} or \code{v_label = 'vowel'}}

\item{utterance_id:}{each unique utterance should have a unique id}

\item{cv_duration:}{the duration of C or V (only the values for vowels will be used)}

\item{utterance_duration:}{the duration of entire utterances}
}
\value{
percent_v: a data frame containing the mean values and standard deviation of \%
}
\description{
This function loads a dataframe as input and returns a dataframe containing the mean values and mean standard deviations of \%V.
}
\details{
\%V is a rhythm metrics based on Ramus, F., Nespor, M., & Mehler, J. (1999). Correlates of linguistic rhythm in the speech signal. Cognition, 73(3), 265-292. It calculates the ratio of vocalic material to the total duration of an utterance.

\verb{\% V: total V duration / total utterance duration}
}
\examples{
df <- data.frame (cv_label  = c("consonant", "vowel", "consonant", "vowel",
                                "consonant", "vowel", "consonant", "vowel"),
                  utterance_id = c("utt_1", "utt_1", "utt_2", "utt_2",
                                   "utt_1", "utt_1", "utt_2", "utt_2"),
                  cv_duration = c(0.1, 0.8, 0.2, 0.5, 0.3, 0.3, 0.4, 0.7),
                  utterance_duration = c(2.4, 2.4, 2.7, 2.7, 2.4, 2.4, 2.7, 2.7))

percentage_v(df, v_label="vowel", utterance_id, cv_duration, utterance_duration)

}
\author{
Cong Zhang, \email{cong.zhang@ru.nl}
}
