% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta_cv.R
\name{plot_delta_cv}
\alias{plot_delta_cv}
\title{Plot Delta C & Delta V}
\usage{
plot_delta_cv(
  df,
  cv_label,
  utterance_id,
  cv_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{cv_label}{column name for segment type (C or V).}

\item{utterance_id}{column name for unique utterance IDs.}

\item{cv_duration}{column name for the duration of C or V.}

\item{save_fig}{default is \code{FALSE}. Change to \code{TRUE} to save the plot.}

\item{fig_path}{default is \code{NULL}. Required if \code{save_fig = TRUE}.}
}
\value{
A boxplot for Delta C and Delta V values.
}
\description{
This function loads a dataframe as input and returns a box plot for Delta C and Delta V values.
}
\examples{
df_test <- data.frame(cv_label = rep(c("c", "v"), 10),
                      utterance_id = rep(paste0("utt_", 1:10), each = 2),
                      cv_duration = runif(20, 0.05, 0.5))

# Not saving the plot
plot_delta_cv(df_test, cv_label, utterance_id, cv_duration, save_fig = FALSE)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
