% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta_cv.R
\name{plot_delta_cv}
\alias{plot_delta_cv}
\title{Plot Delta C & Delta V}
\usage{
plot_delta_cv(
  df,
  cv_label,
  utterance_id,
  cv_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df:}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{cv_label:}{whether a segment is a C(onsonant) or a V(owel)}

\item{utterance_id:}{each unique utterance should have a unique id}

\item{cv_duration:}{the duration of C or V}

\item{save_fig:}{the default value is set to \code{FALSE}. Change to \code{TRUE} to save the plot.}

\item{fig_path:}{the default value is set to \code{NULL}. If \code{save_fig} is set to \code{TRUE}, the path to save the figure is needed.}
}
\value{
A boxplot for Delta C and Delta V values.
}
\description{
This function loads a dataframe as input and returns a box plot for Delta C and Delta V values.
}
\examples{
df <- data.frame (cv_label  = c("c", "v", "c", "v","c", "v", "c", "v"),
                  utterance_id = c("utt_1", "utt_1", "utt_2", "utt_2",
                                   "utt_1", "utt_1", "utt_2", "utt_2"),
                  cv_duration = c(0.1, 0.8, 0.2, 0.5, 0.3, 0.3, 0.4, 0.7))

# Saving the plot
plot_delta_cv(df, cv_label, utterance_id, cv_duration, save_fig=T, fig_path='C:/Users/congzhang/Desktop/')

# Not saving the plot
plot_delta_cv(df, cv_label, utterance_id, cv_duration, save_fig=F)

}
\author{
Cong Zhang, \email{cong.zhang@ru.nl}
}
