% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentage_v.R
\name{plot_percentage_v}
\alias{plot_percentage_v}
\title{Plot Percentage V}
\usage{
plot_percentage_v(
  df,
  cv_label,
  label_name,
  utterance_id,
  cv_duration,
  utterance_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df}{a data frame containing cv_labels, utterance_id, cv_duration, and utterance_duration values.}

\item{cv_label}{column name for the segment labels (e.g., cv_label).}

\item{label_name}{a string to filter the vowels, e.g. \code{label_name = 'vowel'}.}

\item{utterance_id}{column name for unique utterance IDs.}

\item{cv_duration}{column name for the duration of C or V.}

\item{utterance_duration}{column name for the duration of entire utterances.}

\item{save_fig}{default is \code{FALSE}. Change to \code{TRUE} to save the plot.}

\item{fig_path}{default is \code{NULL}. Required if \code{save_fig = TRUE}.}
}
\value{
A boxplot for \%V values.
}
\description{
This function loads a dataframe as input and returns a box plot for \%V values.
}
\examples{
# plot_percentage_v(df, cv_label = cv_label, label_name = "vowel", 
#                  utterance_id = utterance_id, cv_duration = cv_duration, 
#                  utterance_duration = utterance_duration)

}
\author{
Cong Zhang, \email{cong.zhang@newcastle.ac.uk}
}
