% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rpvi_c.R
\name{plot_rpvi}
\alias{plot_rpvi}
\title{Plot rPVI C}
\usage{
plot_rpvi(
  df,
  c_label,
  utterance_id,
  cv_duration,
  save_fig = FALSE,
  fig_path = NULL
)
}
\arguments{
\item{df:}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{c_label:}{a string to filter the consonants, e.g. \code{c_label = 'C'} or \code{c_label = 'consonant'}}

\item{utterance_id:}{each unique utterance should have a unique id}

\item{cv_duration:}{the duration of C or V (only the values for vowels will be used)}
}
\value{
A boxplot for rPVI C values.
}
\description{
This function loads a dataframe as input and returns a box plot for rPVI C values.
}
\examples{
df <- data.frame (cv_label  = c("consonant", "vowel", "consonant", "vowel",
                                "consonant", "vowel", "consonant", "vowel"),
                  utterance_id = c("utt_1", "utt_1", "utt_2", "utt_2",
                                   "utt_1", "utt_1", "utt_2", "utt_2"),
                  cv_duration = c(0.1, 0.8, 0.2, 0.5, 0.3, 0.3, 0.4, 0.7))

# Saving the plot
plot_rpvi(df, c_label="consonant", utterance_id, cv_duration, save_fig=T, fig_path='C:/Users/congzhang/Desktop/')
# Not saving the plot
plot_rpvi(df, c_label="consonant", utterance_id, cv_duration, save_fig=FALSE, fig_path=NULL)

}
\author{
Cong Zhang, \email{cong.zhang@ru.nl}
}
