% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpvi_c.R
\name{rpvi_c}
\alias{rpvi_c}
\title{rPVI C}
\usage{
rpvi_c(df, c_label, utterance_id, cv_duration)
}
\arguments{
\item{df:}{a data frame containing cv_labels, utterance_id, and cv_duration values.}

\item{c_label:}{a string to filter the consonants, e.g. \code{c_label = 'C'} or \code{c_label = 'consonant'}}

\item{utterance_id:}{each unique utterance should have a unique id}

\item{cv_duration:}{the duration of C or V (only the values for vowels will be used)}
}
\value{
rpvi_c: a data frame containing the mean values of rpvi for consonants
}
\description{
This function loads a dataframe as input and returns a dataframe containing the mean values of rPVI of consonants.
}
\details{
rPVI C is a rhythm metrics based on Grabe, E., & Low, E. L. (2002). Durational variability in speech and the rhythm class hypothesis. In Laboratory phonology 7 (pp. 515-546). De Gruyter Mouton.
It calculates the sum of the absolute differences between pairs of consecutive consonantal intervals divided by the number of pairs in the speech sample.
}
\examples{
df <- data.frame (cv_label  = c("consonant", "vowel", "consonant", "vowel",
                                "consonant", "vowel", "consonant", "vowel"),
                  utterance_id = c("utt_1", "utt_1", "utt_2", "utt_2",
                                   "utt_1", "utt_1", "utt_2", "utt_2"),
                  cv_duration = c(0.1, 0.8, 0.2, 0.5, 0.3, 0.3, 0.4, 0.7))

rpvi_c(df, c_label="consonant", utterance_id, cv_duration)

}
\author{
Cong Zhang, \email{cong.zhang@ru.nl}
}
